Moving EZID validation logic into Ezid utility class (DHH-676)
