+++ src/java/edu/ucsd/library/dams/util/EzidException.java
@@ -1,0 +1,10 @@
+package edu.ucsd.library.dams.util;
+
+/**
+ * Embedded exception class.
+**/
+public class EzidException extends Exception {
+    public EzidException( String msg ) {
+        super(msg);
+    }
+}
